Reorganize PromQL endpoint tests into separate files
